almost done 15/10

+++ advocate/lib/perticaularCase.dart
@@ -21,6 +21,8 @@ class _PerticaularCaseState extends State<PerticaularCase> {
       });
       
       await dB.deleteCase(id);
+      // Toast
+      //Navigate Back
 
     }catch(e){
 
@@ -37,7 +39,7 @@ class _PerticaularCaseState extends State<PerticaularCase> {
         wait=true;
       });
 
-      dB.getDetails(thisCase.id).then((value){
+      await dB.getDetails(thisCase.id).then((value){
         thisCase.details=value;
       });
 
@@ -46,6 +48,60 @@ class _PerticaularCaseState extends State<PerticaularCase> {
     }setState(() {
       wait=false;
     });
+  }
+
+  Future<void>deleteDetails(int id)async{
+    try{
+      setState(() {
+        wait=true;
+      });
+      
+      await dB.deleteDetails(id);
+
+    }catch(e){
+
+    }finally{
+      setState(() {
+        wait=false;
+      });
+    }
+  }
+
+  void editCase(){
+    Navigator.pushNamed(context, '/editCase',arguments: {'case':thisCase}); 
+  }
+  
+  void editDetails(int indx){
+    Navigator.pushNamed(context, '/editDetails',arguments: {'details':thisCase.details[indx],'caseId':thisCase.id}); 
+  }
+  
+  void addDate(){
+   Navigator.pushNamed(context, '/addDetails',arguments: {'caseId':thisCase.id}); 
+  }
+
+
+  Future<void>chenageMessagePermission()async{
+    try{
+      setState(() {
+        wait=true;
+      });
+      DbHelper dB=DbHelper();
+      await dB.updateCase(thisCase, thisCase.id);
+
+    }catch(e){
+      print(e);
+    }finally{
+      setState((){
+        wait=false;
+      });
+    }
+  }
+
+  String getDate(String date)
+  {
+    DateTime temp=DateTime.parse(date);
+    String res=temp.day.toString()+"-"+temp.month.toString()+"-"+temp.year.toString();
+    return res;
   }
 
   @override
@@ -64,7 +120,6 @@ class _PerticaularCaseState extends State<PerticaularCase> {
       thisCase=map['case'];
       visited=true;
       getDetails();
-      print("Hello");
     }
 
     return Scaffold(
@@ -78,24 +133,23 @@ class _PerticaularCaseState extends State<PerticaularCase> {
         children: [
           Card(
             child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
               children: <Widget>[
-                CircleAvatar(
-                  backgroundColor: Colors.amberAccent,
-                  child: Text(
-                    thisCase.caseNumber.toString(),
-                  ),
+                Text(
+                  thisCase.clientName+ "  ("+thisCase.clientMobile+")",
                 ),
                 SizedBox(height:10),
-                Text(
-                  "Client Name: "+thisCase.clientName,
-                ),
-                SizedBox(height:10),
-                Text(
-                  "Mobile Number: "+thisCase.clientMobile,
-                ),
-                SizedBox(height:10),
-                Text(
-                  "opponent: "+thisCase.opponent,
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    Text(
+                      "opp.: "+thisCase.opponent,
+                    ),
+                    Text(
+                      "Opp. Adv."+thisCase.opponentAdvocate
+                    ),
+                    SizedBox(width:5),    
+                  ],
                 ),
                 SizedBox(height:10),
                 Text(
@@ -103,35 +157,25 @@ class _PerticaularCaseState extends State<PerticaularCase> {
                 ),
                 SizedBox(height:10),
                 Text(
-                  "Court Name: "+thisCase.courtName,
+                  "Court Name: "+thisCase.courtName+ ", "+thisCase.courtNumber.toString(),
                 ),
                 SizedBox(height:10),
                 Text(
-                  "Court Number: "+thisCase.courtNumber.toString(),
-                ),
-                SizedBox(height:10),
-                Text(
-                  "Case Fee: "+thisCase.caseFee.toString(),
-                ),
-                SizedBox(height:10),
-                Text(
-                  "Registered on: "+thisCase.registeredDate,
-                ),
-                SizedBox(height:10),
-                Text(
-                  "First Date: "+thisCase.firstDate,
+                  "Description: "+thisCase.description
                 ),
 
-                SizedBox(height:30),
+                SizedBox(height:20),
                 
                 Row(
                   children: <Widget>[
                     Text("Message to Client: "),
                     Switch(
-                      value: thisCase.messagePermission,
+                      value: !thisCase.messagePermission,
                       onChanged: (bool value){
                         setState(() {
                           thisCase.messagePermission=value;
+                          // update Case;
+                          chenageMessagePermission();
                         });
                       }
                     )
@@ -143,22 +187,54 @@ class _PerticaularCaseState extends State<PerticaularCase> {
                   children: [
                     MaterialButton(
                       child: Text("Add Date"),
-                      onPressed: (){},
+                      onPressed: (){
+                        addDate();
+                      },
                       color:Colors.green,
                     ),
                     MaterialButton(
                       child: Text("Edit Case"),
-                      onPressed: (){},
+                      onPressed: (){
+                        editCase();
+                      },
                       color:Colors.orange,
                     ),
                     MaterialButton(
                       child: Text("Delete Case"),
                       onPressed: (){
-                        deleteCase(thisCase.id);
+                        showDialog(
+                          context: context,
+                          builder: (context){
+                            return AlertDialog(
+                              title: Text("Dalete Case "+thisCase.caseNumber),
+                              content: Text("All the details of this case will be deleted and you will not be able to reterive it again. Are you sure you want to delete case "+thisCase.caseNumber+"?"),
+                              actions: [
+                                FlatButton(
+                                  onPressed: (){
+                                    Navigator.of(context).pop();
+                                  },
+                                  child: Text("No"),
+                                ),
+                                FlatButton(
+                                  onPressed: (){
+                                    Navigator.of(context).pop();
+                                    deleteCase(thisCase.id);            
+                                  },
+                                  child: Text("Yes"),
+                                ),
+                              ],
+                            );
+                          }
+                        );
                       },
                       color:Colors.red,
                     ),
                   ],
+                ),
+                SizedBox(height:10),
+                OutlineButton(
+                  onPressed: (){},
+                  child: Text("Call for meeting"),
                 ),
               ],
             ),
@@ -173,26 +249,56 @@ class _PerticaularCaseState extends State<PerticaularCase> {
                 child: InkWell(
                   onTap: (){
                     // Go to details Page
+                    Navigator.pushNamed(context, '/page',arguments: {'case':thisCase,"currentPages":index});
                   },
                   child: Column(
                     children:<Widget>[
                       Text(
-                        thisCase.details[index].date
+                        getDate(thisCase.details[index].date)
                       ),
                       Text(
                         thisCase.details[index].stage
                       ),
                       SizedBox(height:30),
                       Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceAround,
                         children: [
                           IconButton(
                             icon: Icon(Icons.delete), 
-                            onPressed: null
+                            onPressed: (){
+                              showDialog(
+                                context: context,
+                                builder: (context){
+                                  return AlertDialog(
+                                    title: Text("Delete Date"),
+                                    content: Text("You will not be able to reterive once you delete this date. Are you sure you want to delete "+thisCase.details[index].date+"?"),
+                                    actions: [
+                                      FlatButton(
+                                        onPressed: (){
+                                          Navigator.of(context).pop();
+                                        },
+                                        child: Text("No"),
+                                      ),
+                                      FlatButton(
+                                        onPressed: (){
+                                          Navigator.of(context).pop();
+                                          // Delete Details
+
+                                        },
+                                        child: Text("Yes"),
+                                      ),
+                                    ],
+                                  );
+                                }
+                              );
+                            }
                           ),
 
                           IconButton(
                             icon: Icon(Icons.edit), 
-                            onPressed: null
+                            onPressed:(){
+                              editDetails(index);
+                            }
                           ),
                         ],
                       ),
